move ValidatePackageManifest to an operator-registry internal/util dir
