+++ internal/pkg/scaffold/olm-catalog/package_manifest_test.go
@@ -24,8 +24,6 @@ import (
 	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold"
 	"github.com/operator-framework/operator-sdk/internal/pkg/scaffold/input"
 	"github.com/operator-framework/operator-sdk/internal/util/diffutil"
-
-	olmregistry "github.com/operator-framework/operator-lifecycle-manager/pkg/controller/registry"
 )
 
 func TestPackageManifest(t *testing.T) {
@@ -70,74 +68,3 @@ const packageManifestExp = `channels:
 defaultChannel: stable
 packageName: app-operator
 `
-
-func TestValidatePackageManifest(t *testing.T) {
-	channels := []olmregistry.PackageChannel{
-		{Name: "foo", CurrentCSVName: "bar"},
-	}
-	pm := &olmregistry.PackageManifest{
-		Channels:           channels,
-		DefaultChannelName: "baz",
-		PackageName:        "test-package",
-	}
-
-	cases := []struct {
-		description string
-		wantErr     bool
-		errMsg      string
-		operation   func(*olmregistry.PackageManifest)
-	}{
-		{
-			"default channel does not exist",
-			true, "default channel baz does not exist in channels", nil,
-		},
-		{
-			"successful validation",
-			false, "",
-			func(pm *olmregistry.PackageManifest) {
-				pm.DefaultChannelName = pm.Channels[0].Name
-			},
-		},
-		{
-			"channels are empty",
-			true, "channels cannot be empty",
-			func(pm *olmregistry.PackageManifest) {
-				pm.Channels = nil
-			},
-		},
-		{
-			"one channel's CSVName is empty",
-			true, "channel foo currentCSV cannot be empty",
-			func(pm *olmregistry.PackageManifest) {
-				pm.Channels = make([]olmregistry.PackageChannel, 1)
-				copy(pm.Channels, channels)
-				pm.Channels[0].CurrentCSVName = ""
-			},
-		},
-		{
-			"duplicate channel name",
-			true, "duplicate package manifest channel name foo; channel names must be unique",
-			func(pm *olmregistry.PackageManifest) {
-				pm.Channels = append(channels, channels...)
-			},
-		},
-	}
-
-	for _, c := range cases {
-		if c.operation != nil {
-			c.operation(pm)
-		}
-		err := ValidatePackageManifest(pm)
-		if c.wantErr {
-			if err == nil {
-				t.Errorf(`%s: expected error "%s", got none`, c.description, c.errMsg)
-			} else if err.Error() != c.errMsg {
-				t.Errorf(`%s: expected error message "%s", got "%s"`, c.description, c.errMsg, err)
-			}
-		} else {
-			if err != nil {
-				t.Errorf(`%s: expected no error, got error "%s"`, c.description, err)
-			}
-		}
-	}
-}
